More partitions for God of partitions
