+++ app/api/api_router.py
@@ -1,1 +1,1 @@
-from app.api.api_routes import users
+from app.api.api_routes import api_routes
