+++ app/api/api_routes/api_routes.py
@@ -1,0 +1,6 @@
+from app.api.api_routes.users import (
+    create_user,
+    get_user,
+    get_users,
+    update_user
+)
