--- app/api/api_routes/users.py
@@ -1,21 +1,0 @@
-from app.api import bp
-
-
-@bp.route('/users/<int:id>', methods=['GET'])
-def get_user(id):
-    pass
-
-
-@bp.route('/users', methods=['GET'])
-def get_users():
-    pass
-
-
-@bp.route('/users', methods=['POST'])
-def create_user():
-    pass
-
-
-@bp.route('/users/<int:id>', methods=['PUT'])
-def update_user(id):
-    pass
