+++ app/api/api_routes/users/create_user.py
@@ -1,0 +1,6 @@
+from app.api import bp
+
+
+@bp.route('/users', methods=['POST'])
+def create_user():
+    pass
