+++ app/api/api_routes/users/get_user.py
@@ -1,0 +1,6 @@
+from app.api import bp
+
+
+@bp.route('/users/<int:id>', methods=['GET'])
+def get_user(id):
+    pass
