+++ app/api/api_routes/users/get_users.py
@@ -1,0 +1,6 @@
+from app.api import bp
+
+
+@bp.route('/users', methods=['GET'])
+def get_users():
+    pass
