+++ app/api/api_routes/users/update_user.py
@@ -1,0 +1,6 @@
+from app.api import bp
+
+
+@bp.route('/users/<int:id>', methods=['PUT'])
+def update_user(id):
+    pass
