+++ app/router.py
@@ -1,0 +1,1 @@
+from app.api import api_router
